🐛 [Base] fix(console): correct line count type and vendor option query
- 【Console】fix type hint for `$lineCount` variable in `DocBlock` command
- 【BidResource】correct vendor options query using `unless` instead of `when` for better readability

+++ app/Console/Commands/DocBlock.php
@@ -81,7 +81,7 @@ class DocBlock extends Command
             $this->info('→ Scanning: ' . $file->getPathname());
             $lines = (array) file($file->getPathname());
             $fileContent = (string) file_get_contents($file->getPathname());
-            $lineCount = (int) count($lines);
+            $lineCount = count($lines);
             $modified = false;
 
             for ($i = 0; $i < $lineCount; $i++) {
